Create ifpermutation_of_string
Check Permutation: Given two strings, write a method to decide if one is a permutation of the
other.
We can also use the definition of a permutation-two words with the same character counts-to implement
this algorithm. We simply iterate through this code, counting how many times each character appears.
Then, afterwards, we compare the two arrays.